[feat]: Moved the client cli api into apps/lib
