+++ apps/src/lib/client/tx.rs
@@ -14,6 +14,7 @@ use namada::ledger::rpc::{TxBroadcastData, TxResponse};
 use namada::ledger::signing::TxSigningKey;
 use namada::ledger::wallet::{Wallet, WalletUtils};
 use namada::ledger::{masp, pos, signing, tx};
+use namada::ledger::queries::Client;
 use namada::proof_of_stake::parameters::PosParams;
 use namada::proto::{Code, Data, Section, Tx};
 use namada::types::address::Address;
@@ -35,7 +36,6 @@ use crate::client::signing::find_pk;
 use crate::client::tx::tx::ProcessTxResponse;
 use crate::config::TendermintMode;
 use crate::facade::tendermint_rpc::endpoint::broadcast::tx_sync::Response;
-use crate::facade::tendermint_rpc::HttpClient;
 use crate::node::ledger::tendermint_node;
 use crate::wallet::{gen_validator_keys, read_and_confirm_encryption_password};
 
@@ -525,8 +525,8 @@ impl masp::ShieldedUtils for CLIShieldedUtils {
     }
 }
 
-pub async fn submit_transfer(
-    client: &HttpClient,
+pub async fn submit_transfer<C: Client + Sync>(
+    client: &C,
     mut ctx: Context,
     args: args::TxTransfer,
 ) -> Result<(), tx::Error> {
